Drastically restrict the set of alternative literal types checked by the solver.
When the solver has attempted to produce a solution using the default
literal type tha has failed, dont' try every type that conforms to
that literal type. Instead, try the bridged class type (which deals
with the common AnyObject case) or one of two other options:
  - For integer literals, try the default floating point type (Double)
  - For string literals, try the standard library's AssertString (this
  is a temporary hack)

This limits exponential blow-up in cases where the literal's type
cannot be determined from context. Addresses rdar://problem/18307267.



Swift SVN r22131

+++ test/Constraints/default_literals.swift
@@ -25,7 +25,7 @@ genericFloatingLiteral(d)
 extension UInt32 {
   func asChar() -> UnicodeScalar { return UnicodeScalar(self) }
 }
-var ch = 65.asChar()
+var ch = UInt32(65).asChar()
 
 // <rdar://problem/14634379>
 extension Int {
